Safari-609.1.11.1/LayoutTests/editing/pasteboard/dom-paste/dom-paste-consecutive-confirmations-expected.txt


Verifies that no menu is shown when the page programmatically triggers paste on a timer after user interaction. To test manually, click or tap the text on the bottom to copy, and then click or tap the editable area above to trigger two programmatic pastes with the menu. Check that permissions for the first programmatic paste do not affect the second programmatic paste, since it is performed on a zero-delay timer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.execCommand('Paste') is true
PASS editor.textContent is "Click here to copy"
PASS document.execCommand('Paste') is true
PASS editor.textContent is "Click here to copy"
PASS successfullyParsed is true

TEST COMPLETE
